mobile: SendSuccess: faster anim, date
